backend/requirements.txt
fastapi
uvicorn
pydantic[dotenv]
pydantic-settings
requests
pyjwt
passlib[bcrypt]
python-multipart
python-dotenv
gitpython
jinja2
cryptography>=42.0.0
textfsm
napalm
